common: initial support for pool sets
The initial support for pool sets includes:
- detection of pool set files and choosing the right path
  (opening/creation of a pool consisting of a single vs. multiple files)
- creation/validation of the headers of all the parts of the pool set
- pool open/create refactoring to allow mapping the usable pool space
  from a single or multiple files

Note that pool set file parser is a subject for another patch.
Currently, there is a dummy util_poolset_parse() function, so in practice,
only the classic pools are supported.  Once the parser is integrated,
the remaining code should (hopefully) work without any modifications.

The refactored pool creation/opening code takes a structure that
describes all the parts of the pool set.  In this implementation it holds
an array of all the files: their paths, sizes, etc.  In case of
a classic pool file, there is just one element in the array.

There are no new unit tests added, as most of the new code is covered by
the existing unit tests, working on a classic pool files.
In practice, the changes should be completely transparent, as for now,
the library works exactly as before.

There is no support for replicas yet.